+++ localization/localization.core.lua
@@ -1,9 +1,0 @@
-local _, addon = ...
-
-local L = setmetatable({}, { __index = function(t, k)
-	local v = tostring(k)
-	rawset(t, k, v)
-	return v
-end})
-
-addon.L = L